requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.1.3
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
email-validator==2.2.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.46
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
fastapi==0.111.0




